add controller to proper handle creating of data
