+++ src/controller/SettingsController.ts
@@ -1,0 +1,22 @@
+import { Request, Response } from "express";
+import { getCustomRepository } from "typeorm";
+import { SettingsRepository } from "../repositories/SettingsRepository";
+
+class SettingsController {
+  async create(request, response) {
+    const { chat, username } = request.body;
+
+    const settingsRepository = getCustomRepository(SettingsRepository);
+  
+    const settings = settingsRepository.create({
+      chat,
+      username,
+    });
+  
+    await settingsRepository.save(settings);
+  
+    return response.json(settings);
+  }
+}
+
+export { SettingsController }
